Move Deck class to its own file
