+++ Assets/Scripts/Core/Deck.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f1627d4a8da68d84c99b22dd9faa624e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
